correctly handle address_width for block placement
Unfortunately, the MTB_BASE register doesn't actually point to the base
of the block, except in the most rigid of cases.  It points to where things
can start from, but the block can be anywhere past that, aligned to the
mask size.  This frees up much more flexibility in placing the MTB buffer,
but at the cost of having to determine the "address width" at runtime to
correctly decode addresses.

Instead, we must _write_ to the position register, read it back and
count bits to determine the actually useful bits. (Thanks ARM)

Without this, decoding would be generally garbage if the MTB buffer
block wasn't actually placed at the very start of the allowed region.

See the Coresight MTB TRM (ARM DDI 0486B), section B1 "Discovery" for
more information.

Signed-off-by: Karl Palsson <karl.palsson@marel.com>